requirements.txt
pyyaml
schema
pymongo
asyncio
enum34==1.1.8
python-telegram-bot